% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{appGUI}
\alias{appGUI}
\title{Graphical User Interface}
\usage{
appGUI()
}
\value{
shiny app
}
\description{
Runs a Graphical User Interface to evaluate the projects for Global CO2 emission analysis app.
}
\details{
By moving the Year slidebar, it will
}
\examples{
\dontrun{
appGUI()
}
}
