% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Data plotting}
\usage{
plotData(tsdata, forecastData = NULL, plotTit, fc)
}
\arguments{
\item{tsdata}{time Series object. Object contains year observations and statistics about CO2 emission worldwide.}

\item{forecastData}{forecasting time series. This data are calculated by forecast. NULL by default}

\item{plotTit}{character string The title of the plot that can be changed by selecting different country uiSelectInput}

\item{fc}{boolean. Passing argument TRUE if user check to checkbox for forecasting, else FALSE}
}
\value{
plotly
}
\description{
Plot data
}
