% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{readFile}
\alias{readFile}
\title{Reading CSV file}
\usage{
readFile()
}
\value{
data.frame
}
\description{
Read CSV file in inst/extdata/co2emission.csv to data.frame variable. This data will be used in this project.
}
