% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{appGUI}
\alias{appGUI}
\title{Graphical User Interface}
\usage{
appGUI()
}
\value{
shiny app
}
\description{
Runs a Graphical User Interface to evaluate the projects for Global CO2 emission analysis app. This app is presented the amount of worldwide CO2 emission of from 1960 to 2020. It is possible to select different countries to see the details. Besides, the prediction for the next few years is also provided. The data set is collected from Data World Bank.
By moving the Year slidebar, it will change the number of observations year by year.
}
\examples{
\dontrun{
appGUI()
}
}
