% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{dataRefactor}
\alias{dataRefactor}
\title{Data refractoring}
\usage{
dataRefactor(dataFile)
}
\arguments{
\item{dataFile}{data.frame Original data file.}
}
\value{
dataFile data.frame Data file after naming the column name
}
\description{
Change current column name into new name for easily manipulated
}
