% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFile.R
\name{dataRefactor}
\alias{dataRefactor}
\title{Change current column name into new name for easily manipulated}
\usage{
dataRefactor(dataFile)
}
\arguments{
\item{dataFile}{data.frame. Original data file.}
}
\value{
data.frame
}
\description{
Change current column name into new name for easily manipulated
}
