% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotData.R
\name{plotData}
\alias{plotData}
\title{Data plotting}
\usage{
plotData(tsdata, forecastData, plotTit)
}
\arguments{
\item{tsdata}{Time Series object. Object contain year observations and statistics about CO2 emission worldwide.}

\item{forecastData}{Forecasting time series. This data are calculated by forecast}

\item{plotTit}{Characters.}
}
\value{
plotly
}
\description{
Plot data
}
